Move gulp utils out of project template
